+++ internal/rewriter/testdata/rewrites/33-union-rowfilter-one-branch/input.sql
@@ -1,0 +1,1 @@
+SELECT * FROM pg.public.orders UNION SELECT * FROM mysql_erp.archive.orders