+++ internal/rewriter/testdata/rewrites/35-mask-aggregate-distinct/input.sql
@@ -1,0 +1,1 @@
+SELECT count(DISTINCT ssn) FROM pg.hr.emp