+++ internal/rewriter/testdata/rewrites/40-explain-rowfilter/input.sql
@@ -1,0 +1,1 @@
+EXPLAIN SELECT * FROM pg.public.orders