+++ internal/rewriter/testdata/rewrites/42-mask-outer-subquery-untouched/input.sql
@@ -1,0 +1,1 @@
+SELECT ssn FROM pg.hr.emp WHERE dept IN (SELECT dept FROM pg.hr.mgr)